Include relationships between tables
This commit now handles parsing relationships
which link tables together.

The syntax for inline relationships is a little
funky as it doesn't match our model of single key
and value for settings. So we have to handle
inline relationships specially and always return
an array of them when one is found.

No effort is made (yet) to validate the correctness
of relationships or to add info the inline refs.